Rename province modal to region selection input
