+++ src/root/components/ThreeWForm/RegionSelectionInput/FilterOutput/styles.module.scss
@@ -47,7 +47,7 @@
         gap: $spacing * 0.5;
 
         .label {
-          font-weight: $base-font-regular;
+          font-weight: $base-font-light;
         }
 
         .dismiss-button {
