+++ src/root/components/ThreeWForm/styles.module.scss
@@ -26,33 +26,12 @@
     color: grey
   }
 
-  .province-button {
+  .region-selection-input {
     flex-grow: initial;
     flex-shrink: 0;
     flex-basis: unset;
     align-self: flex-start;
     margin-top: 1.5rem;
-  }
-
-  .province-modal {
-    .map-description {
-      display: flex;
-      flex-direction: column;
-      gap: $spacing;
-      padding: $spacing;
-      flex-grow: 1;
-    }
-
-    .main-content {
-      position: relative;
-      min-height: 20rem;
-
-      .map {
-        z-index: 0;
-        flex-grow: 1;
-        height: 50rem;
-      }
-    }
   }
 
 }
