+++ src/root/components/ThreeWForm/useThreeWOptions.ts
@@ -138,6 +138,7 @@ export const schema: FormSchema = {
       programme_type: [requiredCondition],
       project_country: [requiredCondition],
       project_districts: [defaultEmptyArrayType],
+      project_admin2: [defaultEmptyArrayType],
       reached_female: [positiveIntegerCondition],
       reached_male: [positiveIntegerCondition],
       reached_other: [positiveIntegerCondition],
@@ -455,6 +456,7 @@ export function transformResponseFieldsToFormFields(projectResponse: Project): F
     programme_type,
     project_country,
     project_districts,
+    project_admin2,
     reached_female,
     reached_male,
     reached_other,
@@ -489,6 +491,7 @@ export function transformResponseFieldsToFormFields(projectResponse: Project): F
     programme_type,
     project_country,
     project_districts,
+    project_admin2,
     reached_female,
     reached_male,
     reached_other,
